Recompile plone and plone-logged-in bundles
